backend/requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
Flask-CORS==4.0.0
